+++ examples/svg/src/main.rs
@@ -26,18 +26,18 @@ fn load_svg(mut commands: Commands, asset_server: ResMut<AssetServer>) {
     // You can also use `VelloSvgBundle`
     commands
         .spawn((
-            VelloSvgHandle(asset_server.load("embedded://svg/assets/fountain.svg")),
+            VelloSvg2d(asset_server.load("embedded://svg/assets/fountain.svg")),
             VelloSvgAnchor::Center,
         ))
         .insert(Transform::from_scale(Vec3::splat(1.0)));
 }
 
-fn rotate(mut svg: Single<&mut Transform, With<VelloSvgHandle>>, time: Res<Time>) {
+fn rotate(mut svg: Single<&mut Transform, With<VelloSvg2d>>, time: Res<Time>) {
     svg.rotate_z(-0.5 * time.delta_secs());
 }
 
 fn gizmos(
-    svg: Single<(&VelloSvgHandle, &GlobalTransform)>,
+    svg: Single<(&VelloSvg2d, &GlobalTransform)>,
     assets: Res<Assets<VelloSvg>>,
     mut gizmos: Gizmos,
 ) {