+++ examples/svg_ui/src/main.rs
@@ -35,6 +35,6 @@ fn load_svg(mut commands: Commands, asset_server: ResMut<AssetServer>) {
             ..default()
         },
         BorderColor::all(css::FUCHSIA.with_alpha(0.5)),
-        VelloSvgHandle(asset_server.load("embedded://svg_ui/assets/fountain.svg")),
+        UiVelloSvg(asset_server.load("embedded://svg_ui/assets/fountain.svg")),
     ));
 }